Better JSON Casting for Rows
This commit is cherry-picked from https://github.com/trinodb/trino/pull/3613/files

In the existing code, rows are converted to JSON arrays. If the row has a lot of fields, then it is really hard to read the JSON output to understand which field each value in the array corresponds to.
Instead, this commit, casts rows as JSON maps with the field name as the key in the map and the field value as the value.

Co-authored-by: Yuya Ebihara <ebyhry@gmail.com>